requirements.txt
# Core deep learning frameworks
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0
tensorflow-lite>=2.13.0

# Computer vision and image processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.0
scikit-image>=0.21.0

# Data handling and scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Machine learning utilities
scikit-learn>=1.3.0
tqdm>=4.65.0
wandb>=0.15.0
tensorboard>=2.13.0

# Model optimization and quantization
onnx>=1.14.0
onnxruntime>=1.15.0
openvino>=2023.1.0

# Development and utilities
jupyter>=1.0.0
ipywidgets>=8.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
argparse>=1.4.0
pathlib2>=2.3.7
requests>=2.31.0

# Dataset downloading
kaggle>=1.5.16

# Cross-platform compatibility
psutil>=5.9.0
platform>=1.0.8

# Web interface for demo (optional)
streamlit>=1.25.0
gradio>=3.40.0
